Major Update
Added basic enemy combat, expanded level design, added finishing flag, comments where added to the code. The intention of this update is to have a playable demo with only one level. It is enough to show how the games run. More update will be done in the future, but as for today, AUG-29-2021, and forward, I'll be making minor changes or adding new feature slowly to solely focus on more advanced projects that I have on mind.